wip routing from reco item to reco item
