+++ frontend-react/src/Main.js
@@ -1,5 +1,37 @@
 import './Main.css'
 
+import {useParams, Link} from 'react-router-dom'
+import {useState, useEffect} from 'react'
+
+const exampleWord = {
+    word: "什么",
+    pronunciation: "shen3 me5",
+    translation: "what"
+}
+
+const numberSuggestions = 20
+
+const exampleSuggestions = []
+
+for (let i=0; i < numberSuggestions; i++) {
+    exampleSuggestions.push(exampleWord)
+}
+
+const baseUrl = "http://explorehsk.com:5000/"
+
+async function getSuggestions(word) {
+    try {
+        const response = await fetch(baseUrl + `query?word=${word}`);
+        const jsonData = await response.json();
+
+        return jsonData;
+
+    } catch(err) {
+        console.log(err)
+    }
+}
+
+
 function Center(props) {
 
     const {word} = props;
@@ -18,10 +50,10 @@ function Center(props) {
 
 
 function Suggestion(props) {
-    const {word, pronunciation, translation, onWordClick} = props;
+    const {word, pronunciation, translation} = props;
 
     return (
-        <div className={"suggestion"} onClick={() => {onWordClick(word)}}>
+        <div className={"suggestion"}>
             <div className={'zh'}>{word}</div>
             <div className={'en meta'}>
                 <div>{pronunciation}</div>
@@ -34,18 +66,19 @@ function Suggestion(props) {
 
 function Suggestions(props) {
 
-    const {words, onWordClick} = props
+    const {words} = props
 
     const wordComponents = words.map(
         word => (
-            <Suggestion 
-                word={word.Word} 
-                pronunciation={word.Pronunciation} 
-                translation={word.Definition}
-                onWordClick={onWordClick}
-                // key={word.word + word.pronunciation} 
-                // TODO: set a unique key, will likely come from the backend
-            />
+            <Link to={`/word/${word.Word}`}>
+                <Suggestion 
+                    word={word.Word} 
+                    pronunciation={word.Pronunciation} 
+                    translation={word.Definition}
+                    // key={word.word + word.pronunciation} 
+                    // TODO: set a unique key, will likely come from the backend
+                />
+            </ Link>
             )
     )
 
@@ -58,15 +91,28 @@ function Suggestions(props) {
 }
 
 
+function Main() {
 
-function Main(props) {
+    const {word} = useParams()
 
-    const {centerWord, suggestionWords, onWordClick} = props
+    const [centerWord, setCenterWord] = useState(exampleWord)
+    const [suggestionWords, setSuggestionWords] = useState(exampleSuggestions)
+
+    function populateFromWord() {
+        const jsonData = getSuggestions(word)        
+        
+        jsonData.then(value => {
+            setCenterWord(value.source)
+            setSuggestionWords(value.most_similar.slice(1)) 
+        })
+    }
+
+    useEffect(populateFromWord, [word])
 
     return (
         <main>
             <Center word={centerWord} />
-            <Suggestions words={suggestionWords} onWordClick={onWordClick}/>
+            <Suggestions words={suggestionWords}/>
         </main>
     )
 }
